Update 7
Added a tag column for each task, so that the user can assign a task to a category.

Added a sort function which allows the user to sort by tag (ascending or descending)

Allowed the user to enter a task by pressing the enter key instead of having to click create task.

Added a pop up error message which appears if the user tries to enter a task without anything inputted in the task field.